requirements.txt
#
# This file is autogenerated by pip-compile with python 3.11
# To update, run:
#
#    make pip-freeze
#
async-timeout==4.0.2
    # via redis
attrs==21.4.0
    # via pytest
cachecontrol==0.12.11
    # via insights-core
certifi==2022.6.15
    # via requests
cffi==1.15.0
    # via
    #   cryptography
    #   pynacl
charset-normalizer==2.0.12
    # via requests
colorlog==4.7.2
    # via -r requirements.in
cryptography==37.0.2
    # via secretstorage
decorator==4.4.2
    # via
    #   networkx
    #   retry
defusedxml==0.7.1
    # via
    #   insights-core
    #   jira
deprecated==1.2.13
    # via
    #   pygithub
    #   redis
fuzzywuzzy==0.18.0
    # via -r requirements.in
idna==3.3
    # via requests
iniconfig==1.1.1
    # via pytest
insights-core==3.0.280
    # via -r requirements.in
jeepney==0.8.0
    # via
    #   keyring
    #   secretstorage
jinja2==3.1.2
    # via insights-core
jira==3.1.1
    # via -r requirements.in
keyring==23.6.0
    # via jira
lockfile==0.12.2
    # via insights-core
markupsafe==2.1.1
    # via jinja2
msgpack==1.0.4
    # via cachecontrol
multi-key-dict==2.0.3
    # via python-jenkins
nestedarchive==0.2.4
    # via -r requirements.in
networkx==2.5.1
    # via -r requirements.in
oauthlib==3.2.1
    # via requests-oauthlib
packaging==21.0
    # via
    #   -r requirements.in
    #   pytest
    #   redis
pbr==5.9.0
    # via python-jenkins
pluggy==1.0.0
    # via pytest
py==1.11.0
    # via
    #   pytest
    #   retry
pycparser==2.21
    # via cffi
pygithub==1.55
    # via -r requirements.in
pyjwt==2.4.0
    # via pygithub
pynacl==1.5.0
    # via pygithub
pyparsing==3.0.9
    # via packaging
pytest==7.1.1
    # via -r requirements.in
python-dateutil==2.8.2
    # via -r requirements.in
python-gitlab==2.10.1
    # via -r requirements.in
python-jenkins==1.7.0
    # via -r requirements.in
python-levenshtein==0.12.2
    # via -r requirements.in
pyyaml==6.0
    # via
    #   -r requirements.in
    #   insights-core
redis==4.3.4
    # via insights-core
requests==2.26.0
    # via
    #   -r requirements.in
    #   cachecontrol
    #   insights-core
    #   jira
    #   nestedarchive
    #   pygithub
    #   python-gitlab
    #   python-jenkins
    #   requests-oauthlib
    #   requests-toolbelt
requests-oauthlib==1.3.1
    # via jira
requests-toolbelt==0.9.1
    # via
    #   jira
    #   python-gitlab
retry==0.9.2
    # via -r requirements.in
ruamel-yaml==0.17.16
    # via -r requirements.in
secretstorage==3.3.2
    # via keyring
six==1.16.0
    # via
    #   insights-core
    #   python-dateutil
    #   python-jenkins
tabulate==0.8.9
    # via -r requirements.in
tomli==2.0.1
    # via pytest
tqdm==4.59.0
    # via -r requirements.in
urllib3==1.26.9
    # via requests
wrapt==1.14.1
    # via deprecated

# The following packages are considered to be unsafe in a requirements file:
# setuptools
